pyproject.toml
[tool.poetry]
name = "amazon-books-coldstart"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
# packages = [{include = "amazon_books_coldstart"}]

[tool.poetry.dependencies]
python = "^3.9"
pre-commit = "^3.3.3"
black = "^23.3.0"
flake8 = "^6.0.0"
isort = "^5.12.0"
numpy = "^1.25.0"
pandas = "^2.0.2"
click = "^8.1.3"
torch = "^2.0.1"
sentence-transformers = "^2.2.2"
faiss-cpu = "^1.7.4"
pytorch-metric-learning = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
